Sonar Commit 2
Second commit for resolving Sonar Cloud issues.

Commit generalizes ID Card component to be usable on the Search Page, as well as the Compare Page.
