+++ src/components/search/components/IDCard/CollectionExt.tsx
@@ -1,0 +1,50 @@
+/* Import Dependencies */
+import { CheckProperty } from 'global/Utilities';
+import { Row, Col } from 'react-bootstrap';
+
+/* Import Types */
+import { Specimen } from 'global/Types';
+
+/* Import Styles */
+import styles from 'components/search/search.module.scss';
+
+/* Import Icons */
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faBoxArchive } from '@fortawesome/free-solid-svg-icons';
+
+
+/* Props Typing */
+interface Props {
+    specimen: Specimen
+};
+
+
+const CollectionExt = (props: Props) => {
+    const { specimen } = props;
+
+    return (
+        <Row className="mt-4">
+            <Col>
+                <h5 className="c-accent">
+                    <FontAwesomeIcon icon={faBoxArchive} />
+                    <span className="ms-1"> Collection </span>
+                </h5>
+
+                <p className={styles.IDCardProperty}>
+                    <span className="fw-bold"> Collecting number: </span>
+                    {CheckProperty(specimen.data['ods:collectingNumber'])}
+                </p>
+                <p className={`${styles.IDCardProperty} mt-2`}>
+                    <span className="fw-bold"> Collector: </span>
+                    {CheckProperty(specimen.data['ods:collector'])}
+                </p>
+                <p className={`${styles.IDCardProperty} mt-2`}>
+                    <span className="fw-bold"> Date collected: </span>
+                    {CheckProperty(specimen.data['ods:dateCollected'])}
+                </p>
+            </Col>
+        </Row>
+    );
+}
+
+export default CollectionExt;
