+++ src/components/search/search.module.scss
@@ -186,7 +186,7 @@
 
 /* ID Card */
 .IDCard {
-    height: Calc(100% - 90px);
+    height: Calc(100% - 40px);
 }
 
 .IDCardIcon {
